backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
python-dotenv
sqlalchemy
psycopg2-binary
numpy
python-multipart
joblib
requests
